fix: crash when deleting/backing up worlds & packs

+++ app/src/main/java/org/levimc/launcher/core/content/ResourcePackManager.java
@@ -126,6 +126,10 @@ public class ResourcePackManager {
     }
 
     public void importPack(Uri packUri, PackOperationCallback callback) {
+        if (executor.isShutdown()) {
+            callback.onError("ResourcePackManager has been shut down");
+            return;
+        }
         executor.execute(() -> {
             try {
                 InputStream inputStream = context.getContentResolver().openInputStream(packUri);
@@ -196,6 +200,10 @@ public class ResourcePackManager {
     }
 
     public void deletePack(ResourcePackItem pack, PackOperationCallback callback) {
+        if (executor.isShutdown()) {
+            callback.onError("ResourcePackManager has been shut down");
+            return;
+        }
         executor.execute(() -> {
             try {
                 if (deleteFile(pack.getFile())) {
